+++ src/App.tsx
@@ -16,6 +16,7 @@ import MyEntry from "./pages/MyEntry.tsx";
 import Scorecards from "./pages/Scorecards.tsx";
 import Groups from "./pages/Groups.tsx";
 import FpiModel from "./pages/FpiModel.tsx";
+import FpiSeason from "./pages/FpiSeason.tsx";
 import { AuthProvider } from "./contexts/AuthContext.tsx";
 
 const queryClient = new QueryClient();
@@ -40,6 +41,7 @@ const App = () => (
             <Route path="/scorecards" element={<Scorecards />} />
             <Route path="/groups" element={<Groups />} />
             <Route path="/models/fpi" element={<FpiModel />} />
+            <Route path="/models/fpi/season" element={<FpiSeason />} />
             {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
             <Route path="*" element={<NotFound />} />
           </Routes>